git: split out IGit and IGitConfiguration files
Split out the IGit and IGitConfiguration interfaces and extension
methods in to their own files.
